Removed unnecessary components
Removed:
-  Function Buttons
-  Resistors associated with the buttons

Moved some components around to make better use of the space on the
board. Decreased the board length and width. Routed the EEPROM traces
and added a bypass capacitor footprint to the power pin of the EEPROM.

Ripped up some traces and instead used polygon pours to route the
signals to make them lower impedance traces. This includes the traces
routed to the headphone jack and the USB VBUS pin.